fixed like comment

+++ src/Controllers/LikeComments.js
@@ -4,9 +4,10 @@ const User = require('../Models/User')
 const Comment = require('../Models/Comment')
 
 const likeCommentPost = async( req, res = response ) => {
-    const {userId, commentId} = req.body;
     try{
-        const user = await User.findById(userId)
+        const { commentId } = req.body;
+        const tokenDecoded = req.tokenDecoded
+        const user = await User.findById(tokenDecoded.id)
         const comment = await Comment.findById(commentId)
 
         if(!user || !comment ){
@@ -15,7 +16,7 @@ const likeCommentPost = async( req, res = response ) => {
             })
         }
         
-        const likedComment = new LikeComment({userId, commentId});
+        const likedComment = new LikeComment({userId: tokenDecoded.id, commentId});
         await likedComment.save()
         
         return res.status(200).json({
@@ -49,9 +50,11 @@ const likeCommentGet = async (req = request,  res = response) =>{
 }
 
 const likeCommentDelete = async (req = request, res = response) => {
-    const likeComment_id = req.params.Id;
     try {
-        const deletedLikeComment = await LikeComment.findOneAndRemove({ commentId: likeComment_id });
+        const likeComment_id = req.params.commentId;
+        const tokenDecoded = req.tokenDecoded;
+
+        const deletedLikeComment = await LikeComment.findOneAndRemove({ commentId: likeComment_id, userId: tokenDecoded.id });
 
         if (!deletedLikeComment) {
             return res.status(404).json({
